chore: move SSIM functions to separate export
It's not a common API and relies on an outdated library that is not easy to bundle.
