+++ src/compare/index.ts
@@ -1,7 +1,5 @@
-export * from './computeDssim.js';
 export * from './computePsnr.js';
 export * from './computeRmse.js';
-export * from './computeSsim.js';
 export * from './subtract.js';
 export * from './add.js';
 export * from './divide.js';
