+++ src/ssim/__tests__/compute_dssim.test.ts
@@ -1,7 +1,7 @@
 import { expect, test } from 'vitest';
 
 import { Image } from '../../Image.js';
-import { computeDssim } from '../computeDssim.js';
+import { computeDssim } from '../compute_dssim.js';
 
 test('twice the same image', () => {
   const image = testUtils.createGreyImage([[5, 5, 5, 10, 10, 10, 15, 15, 15]]);
