+++ src/ssim/compute_ssim.ts
