+++ src/ssim/index.ts
@@ -1,0 +1,2 @@
+export * from './compute_dssim.js';
+export * from './compute_ssim.js';
